Eliminate dependency on File::Copy::Recursive for dircopy().
As of this date, FCR has a bug which prevents it from passing 'make test' and
'make install' on FreeBSD and other important operating systems.  Any CPAN
libraries which depend on FCR will therefore not be reached or installed when
you are using 'cpan' or similar programs on those platforms.

File::Copy::Recursive::Reduced (versions 0.002 and greater) provide drop-in
replacements for the most basic uses of fcopy() and dircopy().

FCR bug report: https://rt.cpan.org/Ticket/Display.html?id=123964